# parent/com.bike.server/target/generated-sources/openapi/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: jacquetq@gmail.com
  description: |
    Bike Rest Api with Spring Boot
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Bike Api
  version: 1.0.0
...
servers:
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/QuentinJacquet/com.bike.api/1.0.0
...
paths:
...
  /station:
    get:
      operationId: getStations
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Station'
                type: array
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/Station'
                type: array
          description: successful operation
      summary: Get all stations
      tags:
      - station
      x-accepts: application/json
      x-tags:
      - tag: station
    post:
      operationId: createStations
      requestBody:
        $ref: '#/components/requestBodies/StationArray'
      responses:
        "200":
          description: Succesful operation
        "405":
          description: Invalid input
      summary: Create stations
      tags:
      - station
      x-contentType: application/json
      x-accepts: application/json
      x-tags:
      - tag: station
  /station/{stationId}:
    delete:
      operationId: deleteStation
      parameters:
      - description: Station id to delete
        explode: false
        in: path
        name: stationId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: Succesful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Station not found
      summary: Deletes a station
      tags:
      - station
      x-accepts: application/json
      x-tags:
      - tag: station
    get:
      description: Returns a single station
      operationId: getStationById
      parameters:
      - description: ID of station to return
        explode: false
        in: path
        name: stationId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Station'
            application/xml:
              schema:
                $ref: '#/components/schemas/Station'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Station not found
      summary: Find station by ID
      tags:
      - station
      x-accepts: application/json
      x-tags:
      - tag: station
  /user:
    post:
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: Created user object
        required: true
      responses:
        default:
          description: successful operation
      summary: Create user
      tags:
      - user
      x-contentType: application/json
      x-accepts: application/json
      x-tags:
      - tag: user
  /user/login:
    get:
      operationId: loginUser
      parameters:
      - description: The user name for login
        explode: true
        in: query
        name: username
        required: true
        schema:
          type: string
        style: form
      - description: The password for login in clear text
        explode: true
        in: query
        name: password
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
            application/xml:
              schema:
                type: string
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              explode: false
              schema:
                format: int32
                type: integer
              style: simple
            X-Expires-After:
              description: date in UTC when token expires
              explode: false
              schema:
                format: date-time
                type: string
              style: simple
        "400":
          description: Invalid username/password supplied
      summary: Logs user into the system
      tags:
      - user
      x-accepts: application/json
      x-tags:
      - tag: user
...
  /user/{username}:
    delete:
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
      - description: The name that needs to be deleted
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Succesful operation
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
      summary: Delete user
      tags:
      - user
      x-accepts: application/json
      x-tags:
      - tag: user
    get:
      operationId: getUserByName
      parameters:
      - description: The name that needs to be fetched. Use user1 for testing.
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
          description: successful operation
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
      summary: Get user by user name
      tags:
      - user
      x-accepts: application/json
      x-tags:
      - tag: user
...
components:
  requestBodies:
...
    StationArray:
      content:
        application/json:
          schema:
            items:
              $ref: '#/components/schemas/Station'
            type: array
      description: List of station object
      required: true
  schemas:
...
    Station:
      example:
        bikes:
        - bikes
        - bikes
        latitude: 6.0274563
        stationName: stationName
        id: 0
        longitude: 1.4658129
        capacity: 5
      properties:
        id:
          format: int64
          readOnly: true
          type: integer
        stationName:
          type: string
        latitude:
          format: float
          type: number
        longitude:
          format: float
          type: number
        capacity:
          description: capacity of station to accept bikes
          format: int32
          type: integer
        bikes:
          items:
            type: string
          type: array
          xml:
            name: bikeIds
            wrapped: true
      type: object
      xml:
        name: Station
...
    User:
      example:
        firstName: firstName
        lastName: lastName
        password: password
        userStatus: 6
        id: 0
        email: email
        username: username
      properties:
        id:
          format: int64
          readOnly: true
          type: integer
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
          writeOnly: true
        userStatus:
          description: User Status
          format: int32
          type: integer
      type: object
      xml:
        name: User